rearange test so it fit with modual
